--- WalletWasabi.Fluent/Controls/DestinationEntry/ViewModels/IMutableAddressHost.cs
@@ -1,8 +1,0 @@
-namespace WalletWasabi.Fluent.Controls.DestinationEntry.ViewModels;
-
-public interface IMutableAddressHost
-{
-    string Text { get; set; }
-    IObservable<string> TextChanged { get; }
-    IObservable<Address?> ParsedAddress { get; }
-}
